Moved DeviceBuilder to RxBleDeviceMock.Builder
