+++ mockrxandroidble/src/main/java/com/polidea/rxandroidble2/mockrxandroidble/RxBleClientMock.java
@@ -61,7 +61,7 @@ public class RxBleClientMock extends RxBleClient {
         /**
          * Add a {@link RxBleDevice} to the mock client.
          *
-         * @param rxBleDevice device that the mocked client should contain. Use {@link DeviceBuilder} to create them.
+         * @param rxBleDevice device that the mocked client should contain. Use {@link RxBleDeviceMock.Builder} to create them.
          */
         public Builder addDevice(@NonNull RxBleDevice rxBleDevice) {
             this.discoverableDevicesSubject.onNext((RxBleDeviceMock) rxBleDevice);
@@ -71,7 +71,7 @@ public class RxBleClientMock extends RxBleClient {
         /**
          * Add a {@link RxBleDevice} to the list of bonded devices.
          *
-         * @param rxBleDevice device that the mocked client should contain. Use {@link DeviceBuilder} to create them.
+         * @param rxBleDevice device that the mocked client should contain. Use {@link RxBleDeviceMock.Builder} to create them.
          */
         public Builder addBondedDevice(@NonNull RxBleDevice rxBleDevice) {
             bondedDevices.add(rxBleDevice);
@@ -86,138 +86,14 @@ public class RxBleClientMock extends RxBleClient {
         }
     }
 
-    public static class DeviceBuilder {
-
-        private int rssi = -1;
-        private String deviceName;
-        private String deviceMacAddress;
-        private byte[] legacyScanRecord;
-        private ScanRecord scanRecord;
-        private RxBleDeviceServices rxBleDeviceServices;
-        private BluetoothDevice bluetoothDevice;
-        private Map<UUID, Observable<byte[]>> characteristicNotificationSources;
-
-        /**
-         * Build a new {@link RxBleDevice}.
-         * <p>
-         * Calling {@link #scanRecord}, {@link #rssi} and {@link #deviceMacAddress}
-         * is required before calling {@link #build()}. All other methods
-         * are optional.
-         */
-        public DeviceBuilder() {
-            this.rxBleDeviceServices = new RxBleDeviceServices(new ArrayList<BluetoothGattService>());
-            this.characteristicNotificationSources = new HashMap<>();
-        }
-
-        /**
-         * Add a {@link BluetoothGattService} to the device. Calling this method is not required.
-         *
-         * @param uuid            service UUID
-         * @param characteristics characteristics that the service should report. Use {@link CharacteristicsBuilder} to create them.
-         */
-        public DeviceBuilder addService(@NonNull UUID uuid, @NonNull List<BluetoothGattCharacteristic> characteristics) {
-            BluetoothGattService bluetoothGattService = new BluetoothGattService(uuid, 0);
-            for (BluetoothGattCharacteristic characteristic : characteristics) {
-                bluetoothGattService.addCharacteristic(characteristic);
-            }
-            rxBleDeviceServices.getBluetoothGattServices().add(bluetoothGattService);
-            return this;
-        }
-
-        /**
-         * Create the {@link RxBleDeviceMock} instance using the configured values.
-         */
-        public RxBleDevice build() {
-            if (this.rssi == -1) throw new IllegalStateException("Rssi is required. DeviceBuilder#rssi should be called.");
-            if (this.deviceMacAddress == null) throw new IllegalStateException("DeviceMacAddress required."
-                    + " DeviceBuilder#deviceMacAddress should be called.");
-            if (this.scanRecord == null && this.legacyScanRecord == null)
-                throw new IllegalStateException("ScanRecord required. DeviceBuilder#scanRecord should be called.");
-            RxBleDeviceMock rxBleDeviceMock;
-            if (scanRecord == null) {
-                 rxBleDeviceMock = new RxBleDeviceMock(deviceName,
-                        deviceMacAddress,
-                        legacyScanRecord,
-                        rssi,
-                        rxBleDeviceServices,
-                        characteristicNotificationSources,
-                        bluetoothDevice);
-            } else {
-                rxBleDeviceMock = new RxBleDeviceMock(deviceName,
-                        deviceMacAddress,
-                        scanRecord,
-                        rssi,
-                        rxBleDeviceServices,
-                        characteristicNotificationSources,
-                        bluetoothDevice);
-            }
-
-            for (BluetoothGattService service : rxBleDeviceServices.getBluetoothGattServices()) {
-                rxBleDeviceMock.addAdvertisedUUID(service.getUuid());
-            }
-            return rxBleDeviceMock;
-        }
-
-        /**
-         * Set a device mac address. Calling this method is required.
-         */
-        public DeviceBuilder deviceMacAddress(@NonNull String deviceMacAddress) {
-            this.deviceMacAddress = deviceMacAddress;
-            return this;
-        }
-
-        /**
-         * Set a device name. Calling this method is not required.
-         */
-        public DeviceBuilder deviceName(@NonNull String deviceName) {
-            this.deviceName = deviceName;
-            return this;
-        }
-
-        /**
-         * Sets a bluetooth device. Calling this method is not required.
-         */
-        public DeviceBuilder bluetoothDevice(@NonNull BluetoothDevice bluetoothDevice) {
-            this.bluetoothDevice = bluetoothDevice;
-            return this;
-        }
-
-        /**
-         * Set an {@link Observable} that will be used to fire characteristic change notifications. It will be subscribed to after
-         * a call to {@link com.polidea.rxandroidble2.RxBleConnection#setupNotification(UUID)}. Calling this method is not required.
-         *
-         * @param characteristicUUID UUID of the characteristic that will be observed for notifications
-         * @param sourceObservable   Observable that will be subscribed to in order to receive characteristic change notifications
-         */
-        public DeviceBuilder notificationSource(@NonNull UUID characteristicUUID, @NonNull Observable<byte[]> sourceObservable) {
-            characteristicNotificationSources.put(characteristicUUID, sourceObservable);
-            return this;
-        }
-
-        /**
-         * Set a rssi that will be reported. Calling this method is required.
-         */
-        public DeviceBuilder rssi(int rssi) {
-            this.rssi = rssi;
-            return this;
-        }
-
-        /**
-         * Set a BLE scan record. Calling either this method or the other {@link #scanRecord(ScanRecord) scanRecord} method is required.
-         */
-        public DeviceBuilder scanRecord(@NonNull byte[] scanRecord) {
-            this.legacyScanRecord = scanRecord;
-            return this;
-        }
-
-        /**
-         * Set a BLE scan record. Calling this method is required.
-         */
-        public DeviceBuilder scanRecord(@NonNull ScanRecord scanRecord) {
-            this.scanRecord = scanRecord;
-            return this;
-        }
-    }
+    /**
+     * Device builder class.
+     * @deprecated Use {@link RxBleDeviceMock.Builder}
+     */
+//    @Deprecated
+//    public static class DeviceBuilder extends RxBleDeviceMock.Builder {
+//
+//    }
 
     public static class CharacteristicsBuilder {
 
@@ -225,7 +101,7 @@ public class RxBleClientMock extends RxBleClient {
 
         /**
          * Build a new {@link BluetoothGattCharacteristic} list.
-         * Should be used in pair with {@link DeviceBuilder#addService}
+         * Should be used in pair with {@link RxBleDeviceMock.Builder#addService}
          */
         public CharacteristicsBuilder() {
             this.bluetoothGattCharacteristics = new ArrayList<>();
